Implementing the basic version of the Parser
This version only handles expressions. Following the rules of Precedence and Associativity, we implement a bottom up approach form the lowest expressions in precedence to the higher one, once we start parsing a token, we start by checking it for the least precedence and going up the precedence tree until we find a suitable place to parse it.