Defining blocks and scopes
A block statement starts with "{", we loop through all statements until we find a closing "}" and store all in-between statements in the block statement.
To execute a block, we need to give it it's own scope, we do this by defining a new environment accessed only by this block of the source code. When we interpret this block, we refer to its local environment instead of the global one.
Environments now are chained from innermost to outmost, this give us the opportunity to access global variables from deeply nested blocks of code